refactor: move PlayerBundle and more player-specific components to willo module
